requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
Flask-Migrate>=4.0
alembic>=1.13
Flask-JWT-Extended>=4.6
passlib[bcrypt]>=1.7
python-dotenv>=1.0
marshmallow>=3.21
Flask-Cors>=4.0
Flask-Limiter>=3.7
redis>=5.0
psycopg[binary]>=3.1
gunicorn>=21.2
python-json-logger>=2.0
bcrypt<4.1

